+++ lib/naver/client.rb
@@ -9,10 +9,11 @@ module Naver # :nodoc:
 
     # Converts the response body to an ObjectifiedHash.
     def self.parse(body)
-      body = body.to_underscore_keys
       if body.is_a?(Hash)
+        body = body.to_underscore_keys
         ObjectifiedHash.new(body)
       elsif body.is_a?(Array)
+        body = body.to_underscore_keys
         body.collect! { |e| ObjectifiedHash.new(e) }
       elsif body
         true